+++ mc/user/dal/dptUserDal.go
@@ -25,7 +25,7 @@ func (d DptUserDal) AfterAction(kind qf.EKind, content interface{}) error {
 //  @param userIds
 //  @return error
 //
-func (d DptUserDal) AddUsers(departId uint, userIds []uint) error {
+func (d DptUserDal) AddUsers(departId uint64, userIds []uint64) error {
 	oldUserIds, err := d.GetUsersByDptId(departId)
 	if err != nil {
 		return err
@@ -51,7 +51,7 @@ func (d DptUserDal) AddUsers(departId uint, userIds []uint) error {
 //  @param userIds
 //  @return error
 //
-func (d DptUserDal) RemoveUser(departId uint, userId uint) error {
+func (d DptUserDal) RemoveUser(departId uint64, userId uint64) error {
 	return d.DB().Where("DepartId = ? AND UserId = ?", departId, userId).Delete(&uModel.DepartUser{}).Error
 }
 
@@ -59,11 +59,11 @@ func (d DptUserDal) RemoveUser(departId uint, userId uint) error {
 // GetUsersByDptId
 //  @Description: 获取部门中所有用户
 //  @param departId
-//  @return []uint
+//  @return []uint64
 //  @return error
 //
-func (d DptUserDal) GetUsersByDptId(departId uint) ([]uint, error) {
-	userIds := make([]uint, 0)
+func (d DptUserDal) GetUsersByDptId(departId uint64) ([]uint64, error) {
+	userIds := make([]uint64, 0)
 	err := d.DB().Where("DepartId = ?", departId).Find(&userIds).Error
 	return userIds, err
 }
@@ -72,11 +72,11 @@ func (d DptUserDal) GetUsersByDptId(departId uint) ([]uint, error) {
 // GetDptsByUserId
 //  @Description: 获取用户所属部门
 //  @param userId
-//  @return []uint
+//  @return []uint64
 //  @return error
 //
-func (d DptUserDal) GetDptsByUserId(userId uint) ([]uint, error) {
-	dptIds := make([]uint, 0)
+func (d DptUserDal) GetDptsByUserId(userId uint64) ([]uint64, error) {
+	dptIds := make([]uint64, 0)
 	err := d.DB().Where("UserId = ?", userId).Find(&dptIds).Error
 	return dptIds, err
 }
