+++ mc/user/userBll.go
@@ -41,10 +41,10 @@ func (u *UserBll) login(ctx *qf.Context) (interface{}, error) {
 	params.LoginId = strings.Replace(params.LoginId, " ", "", -1)
 	params.Password = uUtils.ConvertToMD5([]byte(params.Password))
 	if user, ok := u.userDal.CheckLogin(params.LoginId, params.Password); ok {
-		role, _ := u.userRoleDal.GetUsersByRoleId(uint(user.Id))
-		return GenerateToken(uint(user.Id), role, u.jwtSecret)
+		role, _ := u.userRoleDal.GetUsersByRoleId(user.Id)
+		return GenerateToken(user.Id, role, u.jwtSecret)
 	} else if params.LoginId == devUser.LoginId && params.Password == devUser.Password {
-		return GenerateToken(uint(devUser.Id), []uint{}, u.jwtSecret)
+		return GenerateToken(devUser.Id, []uint64{}, u.jwtSecret)
 	} else {
 		return nil, errors.New("loginId not exist or password error")
 	}
@@ -73,7 +73,7 @@ func (u *UserBll) deleteUser(ctx *qf.Context) (interface{}, error) {
 func (u *UserBll) getUserModel(ctx *qf.Context) (interface{}, error) {
 	var user uModel.User
 	//获取用户角色
-	roles, err := u.userRoleDal.GetRolesByUserId(ctx.UserId)
+	roles, err := u.userRoleDal.GetRolesByUserId(uint64(ctx.UserId))
 	if err != nil {
 		return nil, err
 	}
@@ -105,7 +105,7 @@ func (u *UserBll) getAllUsers(ctx *qf.Context) (interface{}, error) {
 //
 func (u *UserBll) resetPassword(ctx *qf.Context) (interface{}, error) {
 	uId := ctx.GetUIntValue("Id")
-	return nil, u.userDal.SetPassword(uint(uId), uUtils.ConvertToMD5([]byte(defPassword)))
+	return nil, u.userDal.SetPassword(uId, uUtils.ConvertToMD5([]byte(defPassword)))
 }
 
 //
@@ -123,8 +123,8 @@ func (u *UserBll) changePassword(ctx *qf.Context) (interface{}, error) {
 	if err := ctx.Bind(&params); err != nil {
 		return nil, err
 	}
-	if !u.userDal.CheckOldPassword(ctx.UserId, params.OldPassword) {
+	if !u.userDal.CheckOldPassword(uint64(ctx.UserId), params.OldPassword) {
 		return nil, errors.New("old password is incorrect")
 	}
-	return nil, u.userDal.SetPassword(ctx.UserId, uUtils.ConvertToMD5([]byte(params.NewPassword)))
+	return nil, u.userDal.SetPassword(uint64(ctx.UserId), uUtils.ConvertToMD5([]byte(params.NewPassword)))
 }
